refactor graph and plots
