+++ moonstone/plot/__init__.py
@@ -1,1 +1,2 @@
-from .global_plot import PlotCountsStats, PlotMetadataStats  # noqa
+from .counts import PlotCountsStats  # noqa
+from .metadata import PlotMetadataStats  # noqa
