+++ moonstone/plot/graphs/bargraph.py
@@ -1,52 +1,15 @@
-from abc import ABC, abstractmethod
 import math
+from abc import ABC
+
 import numpy as np
 import pandas as pd
 import plotly.graph_objects as go
 import plotly.io
-from typing import Union
 
 from moonstone.plot.utils import (
     check_list_type
 )
-
-
-class BaseGraph(ABC):
-
-    def __init__(self, data: Union[pd.Series, pd.DataFrame], plotting_options: dict = None,
-                 show: bool = True, output_file: Union[bool, str] = False):
-        """
-        :param data: data to plot
-        :param show: set to False if you don't want to show the plot
-        :param output_file: name of the output file
-        :param plotting_options: options of plotting that will override the default setup \n
-                                 [!] Make sure the value given to an argument is of the right type \n
-                                 options allowed : 'log': `bool` ; 'colorbar': `[str, List[str]]` ;
-                                 'tickangle': `[int, float]`
-        """
-        self.data = data
-        self.plotting_options = plotting_options
-
-        if type(show) == bool:
-            self.show = show
-        else:
-            raise ValueError('Error : show value must be a bool, %s given' % type(show).__name__)
-
-        self.output_file = output_file
-        if self.output_file is True:
-            # if no name given for the output file, a generic name is generated
-            if data.name is not None:
-                self.output_file = data.name+"_"+self.__class__.__name__+".html"
-            else:
-                self.output_file = self.__class__.__name__+".html"
-
-    @abstractmethod
-    def plot_one_graph(self, title: str, xlabel: str, ylabel: str):
-        """
-        method that plots the graph
-        needs to be defined in every child class
-        """
-        pass
+from moonstone.plot.graphs.base import BaseGraph
 
 
 class BaseBarGraph(BaseGraph, ABC):
@@ -158,29 +121,3 @@ class DistributionBarGraph(BaseBarGraph):
         if orientation == "v":
             return go.Bar(x=x, y=y, orientation=orientation)
         return go.Bar(x=y, y=x, orientation=orientation)
-
-
-class Histogram(BaseGraph):
-
-    def plot_one_graph(self, title: str, xlabel: str, ylabel: str,
-                       step: Union[int, float]):   # see if we add nbinsx options
-        minimum = int(self.data.min()/step) * step
-        maximum = self.data.max()
-        fig = go.Figure([go.Histogram(x=self.data, xbins=dict(start=minimum, end=maximum, size=step), autobinx=False)])
-
-        if 'log' in self.plotting_options.keys() and self.plotting_options['log'] is True:
-            fig.update_layout(yaxis_type="log")
-        if 'tickangle' in self.plotting_options.keys():
-            fig.update_xaxes(tickangle=self.plotting_options['tickangle'])
-        if 'colorbar' in self.plotting_options.keys():
-            fig.update_traces(marker_color=self.plotting_options['colorbar'])
-
-        fig.update_layout(title_text=title, title_x=0.5)
-        fig.update_xaxes(title_text=xlabel)
-        fig.update_yaxes(title_text=ylabel)
-
-        if self.show is True:
-            fig.show()
-
-        if self.output_file:
-            plotly.io.write_html(fig, self.output_file)
